+++ src/scaffold/index.js
@@ -126,6 +126,23 @@ class Scaffold {
                         projectPackage.devDependencies = this.sortObject(Object.assign(cliPackage.devDependencies, projectPackage.devDependencies));
                         projectPackage.scripts = this.sortObject(Object.assign(cliPackage.scripts, projectPackage.scripts));
 
+                        if (cli.program.angularVersion) {
+                          for(let prop in projectPackage.dependencies) {
+                              if (prop.includes('@angular') && !prop.includes('@angular-devkit')) {
+                                projectPackage.dependencies[prop] = cli.program.angularVersion;
+                              }
+                          }
+                          for(let prop in projectPackage.devDependencies) {
+                            if (prop.includes('@angular') && !prop.includes('@angular-devkit')) {
+                              projectPackage.devDependencies[prop] = cli.program.angularVersion;
+                            }
+                          }
+                          // TODO: figure out ~ > in version
+                          if (cli.program.angularVersion[0] === '7') {
+                            projectPackage.devDependencies['typescript'] = '~3.1.1';
+                          }
+                        }
+
                         if (cli.program.prettier) {
 
                             let prettierDependencies = {